# Minimize buffer copy penalty in SSL read/write.