--- a/common/src/jni/main/cpp/NativeCrypto.cpp
+++ b/common/src/jni/main/cpp/NativeCrypto.cpp
@@ -7367,17 +7367,22 @@ static jint NativeCrypto_SSL_read(JNIEnv* env, jclass, jlong ssl_address, jobjec
                 // TODO: Make sure it's safe to reuse same callbacks repeatedly.
                 temp_ret = sslRead(env, ssl, fdObject, shc, reinterpret_cast<char*>(buf.get()),
                                    chunk_size, sslError, read_timeout_millis);
-                if (temp_ret == THROWN_EXCEPTION && ret > 0) {
+                if ((temp_ret == THROWN_EXCEPTION || temp_ret == -1) && ret > 0) {
+                    // TODO: Should we only check for -1 above?
                     return ret;
                 }
-                if (temp_ret == THROW_SSLEXCEPTION || temp_ret == THROW_SOCKETTIMEOUTEXCEPTION ||
-                    temp_ret == THROWN_EXCEPTION) {
+                if (temp_ret < 0) {
                     // An error was encountered. Handle below.
-                    // TODO: Is it sufficient to check for negative values?
+                    // TODO: Apparently we need to check for all negative values to cover all error
+                    // cases. Is this correct?
                     ret = temp_ret;
                     break;
                 }
                 env->SetByteArrayRegion(b, offset, temp_ret, buf.get());
+                if (env->ExceptionCheck()) {
+                    // Error committing changes to JVM.
+                    return -1;
+                }
                 // Accumulate bytes read.
                 ret += temp_ret;
                 offset += temp_ret;