@bsidhom 2017-02-14
I haven't been able to run the benchmarks on this yet since they're temporarily broken. However, I'd like to get some input on this to understand whether or not this general approach is appropriate. I've left some questions/comments inline as TODOs.