@kruton 2017-02-15
https://github.com/google/conscrypt/pull/92 should fix the benchmarks if you want to cherry-pick that in to test.