@kruton 2017-02-15
I just ran the benchmark and it doesn't seem to help:

```shell
./gradlew jmhJar
java -jar openjdk-benchmarks/build/libs/conscrypt-openjdk-benchmarks-1.0.0-SNAPSHOT-jmh.jar org.conscrypt.benchmarks.ClientSocketThroughputBenchmark.throughput -p messageSize=4096,1048576 -p sslProvider=CONSCRYPT -p socketType=DEFAULT -f 6
```