@bsidhom 2017-02-15
FWIW, the original benchmark didn't seem to show an improvement either. However, I'm able to get massive speedups for large buffers with this.

To benchmark it, I open a local server socket with the `openssl` tool:
```bash
openssl req -x509 -newkey rsa:4096 -nodes -days 365 -subj "/C=US/ST=WA/CN=localhost" -keyout /tmp/key.pem -out /tmp/cert.pem

dd if=/dev/zero bs=1024 count=10485760 | openssl s_server -accept 8888 -cert /tmp/cert.pem -key /tmp/key.pem -no_comp
```

Compile my main test program against Conscrypt:
```java
import java.io.IOException;
import java.io.InputStream;
import java.io.OutputStream;
import java.security.KeyManagementException;
import java.security.NoSuchAlgorithmException;
import java.security.SecureRandom;
import java.security.Security;
import java.security.cert.CertificateException;
import java.security.cert.X509Certificate;
import javax.net.ssl.SSLContext;;
import javax.net.ssl.SSLSocket;
import javax.net.ssl.SSLSocketFactory;
import javax.net.ssl.TrustManager;
import javax.net.ssl.X509TrustManager;
import org.conscrypt.OpenSSLProvider;

public class PullTest {
  private static final String SSL_PROTOCOL = "TLSv1.2";

  private static void copyStream(InputStream input, OutputStream output, int bufferSize) throws IOException {
    byte[] buffer = new byte[bufferSize];
    for (int bytesRead = input.read(buffer); bytesRead != -1; bytesRead = input.read(buffer)) {
      output.write(buffer, 0, bytesRead);
    }
  }

  private static void download(String host, int port, int bufferSize) throws NoSuchAlgorithmException, IOException, KeyManagementException {
    SSLContext context = SSLContext.getInstance(SSL_PROTOCOL);
    context.init(null, new TrustManager[]{new TrustAllTrustManager()}, new SecureRandom());
    SSLSocketFactory socketFactory = context.getSocketFactory();
    try (SSLSocket socket = (SSLSocket) socketFactory.createSocket(host, port)) {
      System.err.println("Using cipher suite: " + socket.getSession().getCipherSuite());
      copyStream(socket.getInputStream(), System.out, bufferSize);
      //copyStream(System.in, socket.getOutputStream(), bufferSize);
    }
  }

  public static void main(String[] args) throws Exception {
    Security.insertProviderAt(new OpenSSLProvider(), 1);
    String host = args[0];
    int port = Integer.parseInt(args[1]);
    int bufferSize = Integer.parseInt(args[2]);
    download(host, port, bufferSize);
  }

  private static final class TrustAllTrustManager implements X509TrustManager {
    @Override
    public void checkClientTrusted(X509Certificate[] chain, String authType)
        throws CertificateException {}

    @Override
    public void checkServerTrusted(X509Certificate[] chain, String authType)
        throws CertificateException {}

    @Override
    public X509Certificate[] getAcceptedIssuers() {
      return new X509Certificate[0];
    }
  }
}
```

And then run it against the version of Conscrypt I'm testing (restarting the server every time):
```bash
time java -cp libconscrypt-master.jar:. PullTest localhost 8888 8192 >/dev/null
time java -cp libconscrypt-master.jar:. PullTest localhost 8888 16777216 >/dev/null
time java -cp libconscrypt-patched.jar:. PullTest localhost 8888 8192 >/dev/null
time java -cp libconscrypt-patched.jar:. PullTest localhost 8888 16777216 >/dev/null
```

On my computer, all of these complete in about 28 seconds (i.e., have a throughput of about 360 MiB/s) except for conscrypt-master with the 16 MiB buffer, which takes over 1.5 hours (averaging about 1.6 MiB/s).

I haven't tested this yet with SSL_write, but I'll update with what I find.