@kruton 2017-02-15
Awesome news. @nmittler is this an indication that our benchmarks for `SSLSocket` are not measuring accurately?