@bsidhom 2017-02-15
For the push test, I ran the same thing in reverse:
```bash
openssl s_server -accept 8889 -cert /tmp/cert.pem -key /tmp/key.pem -no_comp >/dev/null
```

```bash
time dd if=/dev/zero bs=1024 count=10485760 | java -cp libconscrypt-master.jar:. PushTest localhost 8889 8192
time dd if=/dev/zero bs=1024 count=10485760 | java -cp libconscrypt-master.jar:. PushTest localhost 8889 16777216
time dd if=/dev/zero bs=1024 count=10485760 | java -cp libconscrypt-patched.jar:. PushTest localhost 8889 8192
time dd if=/dev/zero bs=1024 count=10485760 | java -cp libconscrypt-patched.jar:. PushTest localhost 8889 16777216
```

The conscrypt-patched tests completed in about 32 seconds (320 MiB/s). The 8 KiB conscrypt-master test was a few seconds slower (290 MiB/s). The 16 MiB conscrypt-master test completed in 884 seconds (12 MiB/s).