@bsidhom 2017-02-15
The results make sense in terms of relative performance since ScopedByteArrayRW is more expensive than ScopedByteArrayRO due to the former committing changes back to the JVM after every JNI call (i.e., twice as many copies).