@nmittler 2017-02-15
@kruton regarding https://github.com/google/conscrypt/pull/93#issuecomment-280107330, it's entirely possible although surprising that our benchmarks are missing this.  It would seem that they're effectively doing the same thing ... pushing a stream of bytes in one direction as fast as possible. I haven't tried this PR locally yet, but it's curious why our new benchmarks don't show the same improvement.