@bsidhom 2017-02-16
There's an issue when the socket is closed with a partially-filled buffer. Fixing now.