@bsidhom 2017-02-16
It looks like sslRead might block if it has no more data to return. If this is the case, we can't easily finagle these repeated delegated sslRead calls into a single InputStream.read()-compatible call without storing additional socket-specific state.