@kruton 2017-02-16
I get this when I run our benchmarks:

```
Iteration  20: Exception in thread "pool-2-thread-1" java.lang.RuntimeException: java.lang.ArrayIndexOutOfBoundsException
	at org.conscrypt.testing.TestServer$ProcessTask.run(TestServer.java:135)
	at java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1142)
	at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:617)
	at java.lang.Thread.run(Thread.java:745)
Caused by: java.lang.ArrayIndexOutOfBoundsException
	at org.conscrypt.NativeCrypto.SSL_read(Native Method)
	at org.conscrypt.OpenSSLSocketImpl$SSLInputStream.read(OpenSSLSocketImpl.java:740)
	at org.conscrypt.testing.TestServer$ProcessTask.readMessage(TestServer.java:143)
	at org.conscrypt.testing.TestServer$ProcessTask.run(TestServer.java:129)
	... 3 more
```

Command I'm using:
```
./gradlew jmhJar
java -jar openjdk-benchmarks/build/libs/conscrypt-openjdk-benchmarks-1.0.0-SNAPSHOT-jmh.jar org.conscrypt.benchmarks.SocketLoopbackThroughputBenchmark.throughput -p messageSize=16777216 -p sslProvider=CONSCRYPT -p cipher=TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256 -f 2
```