@kruton 2017-02-16
I ran the updated benchmarks on this. Just posting my findings. The score is in bytes/sec.

Before:
```
  (messageSize)  (sslProvider)   Mode  Cnt           Score           Error  Units
          16384      CONSCRYPT  thrpt   40  1136631312.888 ±  40366571.672  ops/s
          65536      CONSCRYPT  thrpt   40   731944576.789 ±  25541710.193  ops/s
        1048576      CONSCRYPT  thrpt   40    45798595.010 ±    833215.348  ops/s
        4194304      CONSCRYPT  thrpt   40    11626526.182 ±    996835.745  ops/s
       16777216      CONSCRYPT  thrpt   40     1675864.137 ±   2864496.917  ops/s
```

After:
```
  (messageSize)  (sslProvider)   Mode  Cnt           Score           Error  Units
          16384      CONSCRYPT  thrpt   40  1052094779.596 ±  77330086.388  ops/s
          65536      CONSCRYPT  thrpt   40  1159559845.740 ±  65356599.050  ops/s
        1048576      CONSCRYPT  thrpt   40  1074788096.430 ± 102935173.095  ops/s
        4194304      CONSCRYPT  thrpt   40  1156780087.903 ±  45768763.543  ops/s
       16777216      CONSCRYPT  thrpt   40  1061248002.721 ±  62594978.738  ops/s
```